Rename the "aut" Python package to "autonity_cli"
Make the Python package name consistent with the PyPI package name.

Replace absolute imports within the package with relative imports so
that a renaming like this won't involve so many changes in the future.
