+++ autonity_cli/__version__.py
