+++ pyproject.toml
@@ -29,13 +29,13 @@ Issues = "https://github.com/autonity/autonity-cli/issues"
 Changes = "https://github.com/autonity/autonity-cli/blob/master/CHANGELOG.md"
 
 [project.scripts]
-aut = "aut.__main__:aut"
+aut = "autonity_cli.__main__:aut"
 
 [tool.hatch.version]
-path = "aut/__version__.py"
+path = "autonity_cli/__version__.py"
 
 [tool.hatch.build]
-packages = ["aut"]
+packages = ["autonity_cli"]
 
 [tool.hatch.metadata]
 allow-direct-references = true
@@ -56,7 +56,7 @@ script-tests = [
     "hatch run ./scripts/test_account",
     "hatch run ./scripts/test_token",
 ]
-type-check = "mypy {args:aut tests}"
+type-check = "mypy {args:autonity_cli tests}"
 
 [tool.hatch.envs.github]
 description = "Develop and build against autonity.py from Github"
@@ -83,7 +83,7 @@ all = ["check-code", "check-wheel"]
 check-code = ["ruff check {args:.}", "black --check {args:.}"]
 check-wheel = [
     "hatch build",
-    "check-wheel-contents dist/aut-$(hatch version)-py3-none-any.whl",
+    "check-wheel-contents dist/autonity_cli-$(hatch version)-py3-none-any.whl",
 ]
 format = ["black {args:.}", "ruff check --fix --select I {args:.}"]
